Refactored project structure setup, Removed unused CSS, Updated imports, Added comments
