+++ portfolio/src/App.js
@@ -1,22 +1,26 @@
-import logo from './logo.svg';
-import './App.css';
+// Dependencies
+import React from 'react';
+
+// Styles
+import './assets/styles/App.css';
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
+    <div>
+      <header>
+        <nav>
+          <ul>
+            <li>Home</li>
+            <li>About</li>
+          </ul>
+        </nav>
+        <main>
+          <h1>Welcome to my portfolio!</h1>
+          <p>Here is some information about me.</p>
+        </main>
+        <footer>
+          <p>&copy; 2024</p>
+        </footer>
       </header>
     </div>
   );
