+++ portfolio/src/index.js
@@ -1,8 +1,15 @@
+// Dependencies
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+
+// Styles
+import './assets/styles/index.css';
+
+// Components
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+
+// Utils
+import reportWebVitals from './utils/reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
