+++ portfolio/src/tests/App.test.js
@@ -1,5 +1,11 @@
+// Dependencies
+import React from 'react';
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import { test, expect } from '@jest/globals';
+import '@testing-library/jest-dom/extend-expect'; // for the "toBeInTheDocument" matcher
+
+// Components
+import App from '../App';
 
 test('renders learn react link', () => {
   render(<App />);
